add dryrun option

+++ bin/pagemove.sh
@@ -11,6 +11,7 @@ function show_help(){
     echo "-d DATADIR    - Reuired. Dokuwiki data directory"
     echo "-p PATTERN    - Optional. Sed pattern for page move. Default is empty, no page move"
     echo "-v VERSION    - Optional. Sqlite version, 2 or 3. Default is 3"
+    echo "-t            - Optional. Dry run, do not make changes"
     echo "-h            - show this help"
 }
 
@@ -25,13 +26,15 @@ fi
 datadir=""
 pattern=""
 version=3
+dryruns=0
 
 while [ $# -gt 1 ];do
     case $1 in
         -d) datadir=$2;shift 2;;
         -p) pattern=$2;shift 2;;
         -v) version=$2;shift 2;;
-        -h) show_help;shift 2;;
+        -t) dryruns=1;shift 1;;
+        -h) show_help;shift 1;;
         *) echo "option $1 not recognizable, type -h to see help list";exit;;
     esac
 done
@@ -72,7 +75,9 @@ for old_page in $($sqlite_cmd $sqlite_dbn "select page from entries");do
     #------------------------------------------------
     if [ ! -f $old_fn ];then
         echo -e "\e[00;31m[DELETE]\e[00m $old_page"
-        $sqlite_cmd $sqlite_dbn "delete from entries where pid='$old_md5'"
+        if [ "$dryruns" == "0" ];then
+            $sqlite_cmd $sqlite_dbn "delete from entries where pid='$old_md5'"
+        fi
         continue
     fi
 
@@ -81,13 +86,15 @@ for old_page in $($sqlite_cmd $sqlite_dbn "select page from entries");do
     #------------------------------------------------
     if [ "$old_fn" != "$new_fn" ];then
         echo -e "\e[00;34m[RENAME]\e[00m $old_page \e[00;34m[TO]\e[00m $new_page"
-        #rename filename
-        mkdir -p `dirname $new_fn`
-        mv $old_fn $new_fn
-        #rename pagename
-        $sqlite_cmd $sqlite_dbn "update entries set pid='$new_md5',page='$new_page' where pid='$old_md5'"
-        $sqlite_cmd $sqlite_dbn "update comments set pid='$new_md5' where pid='$old_md5'"
-        $sqlite_cmd $sqlite_dbn "update tags set pid='$new_md5' where pid='$old_md5'"
-        $sqlite_cmd $sqlite_dbn "update subscriptions set pid='$new_md5' where pid='$old_md5'"
+        if [ "$dryruns" == "0" ];then
+            #rename filename
+            mkdir -p `dirname $new_fn`
+            mv $old_fn $new_fn
+            #rename pagename
+            $sqlite_cmd $sqlite_dbn "update entries set pid='$new_md5',page='$new_page' where pid='$old_md5'"
+            $sqlite_cmd $sqlite_dbn "update comments set pid='$new_md5' where pid='$old_md5'"
+            $sqlite_cmd $sqlite_dbn "update tags set pid='$new_md5' where pid='$old_md5'"
+            $sqlite_cmd $sqlite_dbn "update subscriptions set pid='$new_md5' where pid='$old_md5'"
+        fi
     fi
 done
